[FIX] merge 오류 수정 - jwt.yml 중복 설정 제거 - domain.auth -> global.oauth 로 수정
