+++ src/main/resources/application-jwt.yml
@@ -6,15 +6,9 @@ jwt:
   refresh:
     expiration: 600000      # 초 단위, 약 일주일
     header: Authorization
-
-  refresh:
-    expiration: 600000      # 초 단위, 약 일주일
-    header: Authorization
-
   google:
     access:
       expiration: 3600      # 초 단위, 60min
-
     refresh:
       expiration: 600000      # 초 단위, 약 일주일
       header: Authorization
